NIFI-5673 Addressing review feedback
